Add Requests, Controller, Services, Repositories, Migrations, Models. Improvements on architecture and docker

+++ database/migrations/2019_04_30_115707_create_shelters_table.php
@@ -17,7 +17,7 @@ class CreateSheltersTable extends Migration
             $table->bigIncrements('id');
             $table->string('name');
             $table->string('description');
-            $table->integer('cnpj');
+            $table->bigInteger('cnpj');
             $table->timestamps();
         });
     }
